renamed files containing old ai_design name
