--- src/vai_lab/Core/tests/test_ai_design_core.py
@@ -1,15 +1,0 @@
-"""
-Tests for vai_lab.Core.AI_design_core.py
-"""
-
-from vai_lab.Core import AI_design_core
-
-def test_core_init():
-
-    core = AI_design_core.Core()
-
-    assert core
-
-
-
-
